refactor(frontend): Create Alert component
